+++ modules/fx_trace/lib/src/emitter/fx_event.dart
@@ -1,7 +1,14 @@
+import 'dart:async';
+
 import 'fx_emitter.dart';
 
 class FxEvent {
   const FxEvent();
 
   void emit() => FxEmitter().emit(this);
+}
+
+class AsyncFxEvent<T> extends FxEvent{
+  final Completer<T> task;
+  AsyncFxEvent(this.task);
 }