+++ modules/fx_trace/lib/src/mixin.dart
@@ -1,4 +1,5 @@
 import 'model/model.dart';
+import 'trace/trace.dart';
 
 typedef ExceptionCallback = void Function(Trace trace);
 
@@ -6,6 +7,8 @@ class FxTrace with TraceMixin {
   FxTrace._();
 
   static FxTrace? _instance;
+
+  static LogLevel minLogLevel = LogLevel.info;
 
   factory FxTrace() {
     _instance ??= FxTrace._();
@@ -24,8 +27,6 @@ mixin TraceMixin {
     _actions.add(listener);
   }
 
-  LogLevel get logLevel => LogLevel.debug;
-
   void removeTraceListener(ExceptionCallback listener) {
     _actions.remove(listener);
   }
@@ -36,12 +37,16 @@ mixin TraceMixin {
 
   void notifyTrace(Trace trace) {
     if (trace is LogTrace) {
-      bool isLowLever = trace.level.index < logLevel.index;
-      if (isLowLever) return;
+      bool isLowLevel = trace.level.index < FxTrace.minLogLevel.index;
+      if (isLowLevel) return;
     }
 
     for (ExceptionCallback action in _actions) {
-      action(trace);
+      try {
+        action(trace);
+      } catch (e) {
+        // 防止单个监听器异常影响其他监听器
+      }
     }
   }
 }