+++ modules/publish.dart
@@ -6,13 +6,15 @@ void main() async {
   // String module  = 'fx_env';
   // String module  = 'fx_platform_adapter';
   // String module  = 'fx_string';
-  // String module  = 'fx_trace';
+  // String module = 'fx_trace';
   // String module  = 'fx_go_router_ext';
-  String module  = 'fx_go_router_ext';
+  // String module = 'fx_string';
+  // String module = 'fx_dio';
+  String module = 'fx_dao';
   publishModule(module);
 }
 
-Future<void> publishModule(String name) async{
+Future<void> publishModule(String name) async {
   Directory dir = Directory.current;
   String path = p.join(dir.path, 'modules', name);
   print(path);