+++ modules/fx_dio/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.4
 
+- 0.0.4+3: paginate
 - 0.0.4+2: port can be null
 - detail error log