+++ modules/fx_string/CHANGELOG.md
@@ -1,7 +1,13 @@
+## 0.0.3
+
+* support: [.color]/[toHex]
+
+## 0.0.2
+
+* support: [checkMarkdown]/[isMarkdown]
+
 ## 0.0.1
 
 * support: [separator]
 
-## 0.0.2 
 
-* support: [checkMarkdown]/[isMarkdown]